refactor: change location of file with utils functions
