+++ client/src/utils/profilePageUtils.ts
